refactor(fbuild-packages): split disk_cache/index.rs into submodules to satisfy LOC gate
