+++ crates/fbuild-packages/src/disk_cache/index/pid.rs
@@ -1,0 +1,36 @@
+//! Platform-specific PID liveness check used by lease reaping.
+
+/// Check if a PID is alive. Platform-specific.
+pub(super) fn is_pid_alive(pid: u32) -> bool {
+    #[cfg(unix)]
+    {
+        // kill(pid, 0) checks if process exists without sending a signal.
+        // Use raw FFI to avoid a libc crate dependency (matters for musl builds).
+        extern "C" {
+            fn kill(pid: i32, sig: i32) -> i32;
+        }
+        unsafe { kill(pid as i32, 0) == 0 }
+    }
+    #[cfg(windows)]
+    {
+        // Use OpenProcess to check if PID is alive (fast, no subprocess).
+        // PROCESS_QUERY_LIMITED_INFORMATION = 0x1000
+        extern "system" {
+            fn OpenProcess(access: u32, inherit: i32, pid: u32) -> *mut std::ffi::c_void;
+            fn CloseHandle(handle: *mut std::ffi::c_void) -> i32;
+        }
+        const PROCESS_QUERY_LIMITED_INFORMATION: u32 = 0x1000;
+        let handle = unsafe { OpenProcess(PROCESS_QUERY_LIMITED_INFORMATION, 0, pid) };
+        if handle.is_null() {
+            false
+        } else {
+            unsafe { CloseHandle(handle) };
+            true
+        }
+    }
+    #[cfg(not(any(unix, windows)))]
+    {
+        let _ = pid;
+        false
+    }
+}
